v3.0.0: User metrics, expanded type labels, legacy metric removal
